backend/requirements.txt
# Python dependencies
firebase-admin==6.1.0
RPi.GPIO==0.7.0
pyfingerprint==0.5
face-recognition==1.3.0
opencv-python==4.5.5.64
numpy==1.22.4
smbus2==0.4.1
python-dotenv==0.21.0
pickle5==0.0.12

# For React frontend (package.json will handle these)
# react==17.0.2
# react-dom==17.0.2
# firebase==9.9.0
# antd==4.21.0
# react-router-dom==5.3.0